flask_agent_server2/tax_advisor_agent/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "tax_advisor_agent"
version = "0.1.0"
description = "Professional Tax Advisor - AI-powered Tax Planning and Optimization System"
authors = ["Google LLC"]
readme = "README.md"
packages = [{include = "tax_advisor_agent"}]

[tool.poetry.dependencies]
python = "^3.10"
google-adk = "^0.1.0"
google-cloud-aiplatform = {extras = ["reasoningengine"], version = "^1.68.0"}
google-genai = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
google-cloud-aiplatform = {extras = ["reasoningengine"], version = "^1.68.0"}

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
target-version = ['py310'] 